refactor(ai_handler): move streaming response handling and Azure token generation to helpers
